Break Development into modules
Why:

The Development class was getting larger than comfortable. Since it's
not apparent that we're missing models or service objects (though the
latter seems probably), and since validations are about to get more
complex, it seemed like the right time

This change addresses the need by:

* Breaking Development into modules based on function: callbacks,
validation, location, etc. Some use the ActiveSupport::Concern interface
to hook in, while others are normal modules.

* Naming Callbacks something other than Callbacks -- it seems like this
module wasn't loading entirely due to naming conflicts -- something
else named 'Callbacks' was being loaded.

* Updating test to check new validation -- tothu and commsf required.
